Add basic dependencies
Roles for rbenv and nginx

Ruby 2.3.6 requires openssl1.0, Ubuntu Bionic has openssl1.1 by default

The VM needs more RAM to compile ruby. Maybe not 2G worth, but that is
what I am giving it.

Deploy right now is simply copy. The "api" is to call the deploy.yml,
which will do the rest. Should be able to swap out for ansitrano later